Add libpcap-style loop function
This will allow greater performance as it will allow a reusable record
pointer to be used instead of flowtuple_get_next()'s method of giving
you a fresh flowtuple_record_t pointer that has to be freed each time.

Also includes changes to flow2ascii.c to work with changes.